Version 1.1.8, including a funny bug fix
So yeah, some minor cleanups of removing helper functions and directly inlining them (phasescale and mvv-lva) and a small code change from an if else block to switch case inside makemove. For the bug, i had endgames and openings / middlegames the wrong way. So what did this cause? Well, Null move pruning didnt fire in opening and middlegame, but it did in endgame, which is completely backwards. Also, delta pruning was affected.